refactor: move vehicle domain position
