+++ src/main/java/racingcar/domain/circuit/vehicle/Vehicle.java
@@ -1,4 +1,4 @@
-package racingcar.domain.circuit;
+package racingcar.domain.circuit.vehicle;
 
 public interface Vehicle {
