+++ src/test/java/racingcar/domain/circuit/CarPositionTest.java
@@ -1,8 +1,8 @@
 package racingcar.domain.circuit;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.circuit.vehicle.CarPosition;
 
 import static org.junit.jupiter.api.Assertions.*;
